+++ test/test_loads.py
@@ -1,12 +1,11 @@
 from   __future__     import unicode_literals
-import sys
 import pytest
-from   javaproperties import loads
+from   javaproperties import InvalidUEscapeError, loads
 
-need_ordereddict = pytest.mark.skipif(
-    sys.version_info[:2] < (2,7) or sys.version_info[:2] == (3,0),
-    reason='No OrderedDict before 2.7/3.1',
-)
+try:
+    from collections import OrderedDict
+except ImportError:
+    from ordereddict import OrderedDict
 
 def test_loads_simple():
     assert loads('key=value') == {"key": "value"}
@@ -170,15 +169,11 @@ def test_loads_multiple_crlf():
 def test_loads_multiple_cr():
     assert loads('key = value\rfoo = bar') == {"key": "value", "foo": "bar"}
 
-@need_ordereddict
 def test_loads_multiple_ordereddict():
-    from collections import OrderedDict
     assert loads('key = value\nfoo = bar', object_pairs_hook=OrderedDict) == \
         OrderedDict([("key", "value"), ("foo", "bar")])
 
-@need_ordereddict
 def test_loads_multiple_ordereddict_rev():
-    from collections import OrderedDict
     assert loads('foo = bar\nkey = value', object_pairs_hook=OrderedDict) == \
         OrderedDict([("foo", "bar"), ("key", "value")])
 
@@ -314,3 +309,127 @@ def test_loads_form_feed_separator():
 
 def test_loads_escaped_form_feed():
     assert loads('key\\\fvalue=pair') == {"key\fvalue": "pair"}
+
+def test_loads_u_hex_invalid():
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('bad = \\uabcx')
+    assert excinfo.value.escape == '\\uabcx'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\uabcx'
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('\\uabcx = bad')
+    assert excinfo.value.escape == '\\uabcx'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\uabcx'
+
+def test_loads_u_hex_space():
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('bad = \\uabc ')
+    assert excinfo.value.escape == '\\uabc '
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\uabc '
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('\\uabc  = bad')
+    assert excinfo.value.escape == '\\uabc'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\uabc'
+
+def test_loads_u_hex_eof():
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('bad = \\uabc')
+    assert excinfo.value.escape == '\\uabc'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\uabc'
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('\\uabc = bad')
+    assert excinfo.value.escape == '\\uabc'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\uabc'
+
+def test_loads_u_invalid_hex():
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('bad = \\uxabc')
+    assert excinfo.value.escape == '\\uxabc'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\uxabc'
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('\\uxabc = bad')
+    assert excinfo.value.escape == '\\uxabc'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\uxabc'
+
+def test_loads_u_space_hex():
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('bad = \\u abcx')
+    assert excinfo.value.escape == '\\u abc'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\u abc'
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('\\u abcx = bad')
+    assert excinfo.value.escape == '\\u'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\u'
+
+def test_loads_u_eof():
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('bad = \\u')
+    assert excinfo.value.escape == '\\u'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\u'
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('\\u')
+    assert excinfo.value.escape == '\\u'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\u'
+
+def test_loads_u_invalid_eof():
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('bad = \\ux')
+    assert excinfo.value.escape == '\\ux'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\ux'
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('\\ux = bad')
+    assert excinfo.value.escape == '\\ux'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\ux'
+
+def test_loads_u_space_eof():
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('bad = \\u ')
+    assert excinfo.value.escape == '\\u '
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\u '
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('\\u  = bad')
+    assert excinfo.value.escape == '\\u'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\u'
+
+def test_loads_u_hex_sep():
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('\\uab=bad')
+    assert excinfo.value.escape == '\\uab'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\uab'
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('\\uab:bad')
+    assert excinfo.value.escape == '\\uab'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\uab'
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('\\uab bad')
+    assert excinfo.value.escape == '\\uab'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\uab'
+
+def test_loads_u_escaped_escape():
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('bad = \\uab\\cd')
+    assert excinfo.value.escape == '\\uab\\c'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\uab\\c'
+
+def test_loads_u_u_escape_escape():
+    with pytest.raises(InvalidUEscapeError) as excinfo:
+        loads('bad = \\uab\\u0063d')
+    assert excinfo.value.escape == '\\uab\\u'
+    assert str(excinfo.value) == 'Invalid \\u escape sequence: \\uab\\u'
+
+def test_loads_u_extra():
+    assert loads('the = \\u00f0e') == {"the": "\xF0e"}
+    assert loads('\\u00f0e = the') == {"\xF0e": "the"}
+
+def test_loads_big_U_escape():
+    assert loads('goat = \\U0001F410') == {"goat": "U0001F410"}
+
+def test_loads_key_unicode_eq_value():
+    assert loads('key\\u003Dvalue') == {"key=value": ""}
+
+def test_loads_key_unicode_colon_value():
+    assert loads('key\\u003Avalue') == {"key:value": ""}
+
+def test_loads_key_unicode_space_value():
+    assert loads('key\\u0020value') == {"key value": ""}
+
+# escaped backslash + uXXXX